VolumeRegistration

VolumeRegistration.find_deformation_map — Method

Find deformation maps by splitting the dataset in blocks and aligning blocks with subpixel precision

Arguments

moving: the stack to be registered

reference: the stack to be registered to

border_σ: how far to fade out the borders of the whole image/volume

`block_size::NTuple{N, Integer}: size of blocks to compute deformations

block_border_σ::Union{Real, NTuple{N, Real}}: how far to fade out the

max_shift::Union{Real, NTuple{N, Integer}}: maximum displacement of a

a block in each dimension

σ_filter: low-pass filter width

upsampling: upsampling of the registration for subpixel alignment

upsample_padding: how far to pad from the local maximum for upsampling

snr_n_smooths::Integer=2: number

of times the correlation matrices are smoothed by wieghting with adjacent ones (if they are under signal-to-noise ratio)

snr_threshold::Real: the threshold of the "peakiness" of the correlation matrix, if it's smaller than that, it's value is

obtained by smoothing neighbors

snr_n_pad::Integer: window size of the signal-to-noise calculation

VolumeRegistration.find_translation — Method

Find the shift to move moving by to align with reference

Arguments

border_σ: the width of the border kernel for a smooth faloff towards the edges

upsampling::Integer=1: If bigger than 1, how much to upsample the shifts by (for subpixel registration)

VolumeRegistration.make_reference — Method

make_reference(stack; kwargs...)

Make a reference for a stack

Arguments

stack: the data to be aligned, time is assumed to be the list dimension

ref_win

VolumeRegistration.translate — Method

Translate an image by known shift using either reindexing (for integer shifts) of FFT-based phase-space translation for non-integer shifts.

VolumeRegistration.warp_nonrigid — Method

Corrects a sequence of imaging stacks with a transformation found through non-rigid registration

VolumeRegistration.warp_nonrigid — Method

Corrects a plane or volume with a transformation found through non-rigid registration

VolumeRegistration.KrigingUpsampler — Method

Construct an upsampler with its matrix

VolumeRegistration.calc_snr — Method

Computes the signal to noise ratio of a phase correlation array, defined as the ratio of the maximum phase correlation in the window divided by the maximal phase correlation outside of the region where the maximum is (the idea being that it will be low if the phase correlation is flat versus peaked)

VolumeRegistration.extract_low_frequencies — Method

Takes the data corresponding to the real part of the corners of the phase correlation array of interest for shift finding

extract_low_frequencies(data, corner_size)
extract_low_frequencies(data, corner_size, interpolate_middle)

Arugments

VolumeRegistration.fft_translate — Method

Translate using discrete fourier transform

VolumeRegistration.gaussian_fft_filter — Method

Gaussian filter in the fourier domain (a Gaussian in the fourier domain is again a gaussian with inverse variance)

VolumeRegistration.prepare_deformation_map_calc — Method

Prepare everything common for calculating deformation maps

VolumeRegistration.refine_reference — Method

Iterative refinement of a reference image by aligning more frames and then making a new reference out of the most correlated, aligned frames

VolumeRegistration.split_into_blocks — Method

Splits data into an array of blocks, augmenting the last blocks which may not be full to keep consistent size. Cast immediately into complex for further processing

VolumeRegistration.to_ntuple — Method

Small utilities to get window size from iterables or single number

VolumeRegistration.upsampled_shift — Method

Finds the shift in the area around the center of the orignial window with upsampling