VolumeRegistration

VolumeRegistration.KrigingUpsampler

VolumeRegistration.calc_snr

VolumeRegistration.extract_low_frequencies

VolumeRegistration.fft_translate

VolumeRegistration.find_translation

VolumeRegistration.gaussian_fft_filter

VolumeRegistration.split_into_blocks

VolumeRegistration.to_ntuple

VolumeRegistration.upsampled_shift

VolumeRegistration.find_translation — Method

Find the shift to move moving by to align with reference

Arguments

border_σ: the width of the border kernel for a smooth faloff towards the edges

upsampling::Integer=1: If bigger than 1, how much to upsample the shifts by (for subpixel registration)

VolumeRegistration.KrigingUpsampler — Method

Construct an upsampler with its matrix

VolumeRegistration.calc_snr — Method

Computes the signal to noise ratio of a phase correlation array, defined as the ratio of the maximum phase correlation in the window divided by the maximal phase correlation outside of the region where the maximum is (the idea being that it will be low if the phase correlation is flat versus peaked)

VolumeRegistration.extract_low_frequencies — Method

Takes the data corresponding to the real part of the corners of the phase correlation array of interest for shift finding

extract_low_frequencies(data, corner_size)
extract_low_frequencies(data, corner_size, interpolate_middle)

Arugments

VolumeRegistration.fft_translate — Method

Translate using discrete fourier transform

VolumeRegistration.gaussian_fft_filter — Method

Gaussian filter in the fourier domain (a Gaussian in the fourier domain is again a gaussian with inverse variance)

VolumeRegistration.split_into_blocks — Method

Splits data into an array of blocks, augmenting the last blocks which may not be full to keep consistent size. Cast immediately into complex for further processing

VolumeRegistration.to_ntuple — Method

Small utilities to get window size from iterables or single number

VolumeRegistration.upsampled_shift — Method

Finds the shift in the area around the center of the orignial window with upsampling